@@ -1,45 +1,43 @@
 {
     "name": "damienalexandre/jolitypo-website",
     "description": "Demo website for jolicode/jolitypo",
+    "type": "project",
+    "license": "MIT",
     "authors": [
         {
             "name": "Damien Alexandre",
             "email": "dalexandre@jolicode.com"
         }
     ],
-    "type": "project",
-    "license": "MIT",
     "require": {
         "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "jolicode/jolitypo": "^1.0",
-        "sensio/framework-extra-bundle": "^5.6",
-        "symfony/apache-pack": "^1.0",
-        "symfony/browser-kit": "5.4.*",
-        "symfony/console": "5.4.*",
-        "symfony/css-selector": "5.4.*",
-        "symfony/dotenv": "5.4.*",
+        "symfony/browser-kit": "6.2.*",
+        "symfony/console": "6.2.*",
+        "symfony/css-selector": "6.2.*",
+        "symfony/dotenv": "6.2.*",
         "symfony/flex": "^2",
-        "symfony/form": "5.4.*",
-        "symfony/framework-bundle": "5.4.*",
-        "symfony/intl": "5.4.*",
+        "symfony/form": "6.2.*",
+        "symfony/framework-bundle": "6.2.*",
+        "symfony/intl": "6.2.*",
         "symfony/monolog-bundle": "^3.0",
-        "symfony/runtime": "5.4.*",
-        "symfony/twig-bundle": "5.4.*",
-        "symfony/validator": "5.4.*",
+        "symfony/runtime": "6.2.*",
+        "symfony/twig-bundle": "6.2.*",
+        "symfony/validator": "6.2.*",
         "symfony/webpack-encore-bundle": "^1.8",
-        "symfony/yaml": "5.4.*",
+        "symfony/yaml": "6.2.*",
         "twig/extra-bundle": "^3.0",
         "twig/twig": "^3.0"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "^5.3",
-        "symfony/debug-bundle": "5.4.*",
+        "symfony/phpunit-bridge": "6.2.*",
+        "symfony/debug-bundle": "6.2.*",
         "symfony/maker-bundle": "^1.23",
-        "symfony/stopwatch": "5.4.*",
-        "symfony/var-dumper": "5.4.*",
-        "symfony/web-profiler-bundle": "5.4.*"
+        "symfony/stopwatch": "6.2.*",
+        "symfony/var-dumper": "6.2.*",
+        "symfony/web-profiler-bundle": "6.2.*"
     },
     "config": {
         "optimize-autoloader": true,
@@ -93,7 +91,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": true,
-            "require": "5.4.*"
+            "require": "6.2.*"
         }
     }
 }
