@@ -10,9 +10,8 @@ class HomepageControllerTest extends WebTestCase
     {
         $client = static::createClient();
         $client->request('GET', '/');
-        $response = $client->getResponse();
 
-        self::assertSame(200, $response->getStatusCode(), $response->getContent());
+        self::assertResponseIsSuccessful();
     }
 
     public function testSubmitWorks()
@@ -26,9 +25,8 @@ class HomepageControllerTest extends WebTestCase
                 'content' => 'Hello Damien !',
             ]
         ]);
-        $response = $client->getResponse();
 
-        self::assertSame(200, $response->getStatusCode(), $response->getContent());
+        self::assertResponseIsSuccessful();
         self::assertSame(1, $crawler->filter('textarea.form__result-client')->count());
         self::assertSame(1, $crawler->filter('textarea.form__result-html')->count());
     }
@@ -44,9 +42,8 @@ class HomepageControllerTest extends WebTestCase
                 'content' => 'Hello Damien !',
             ]
         ]);
-        $response = $client->getResponse();
 
-        self::assertSame(200, $response->getStatusCode(), $response->getContent());
+        self::assertResponseIsSuccessful();
         self::assertSame(0, $crawler->filter('textarea.form__result-client')->count());
         self::assertSame(0, $crawler->filter('textarea.form__result-html')->count());
         self::assertSame(1, $crawler->filter('html div.error-message', 'This value should not be null.')->count());
@@ -63,9 +60,8 @@ class HomepageControllerTest extends WebTestCase
                 'content' => '',
             ]
         ]);
-        $response = $client->getResponse();
 
-        self::assertSame(200, $response->getStatusCode(), $response->getContent());
+        self::assertResponseIsSuccessful();
         self::assertSame(0, $crawler->filter('textarea.form__result-client')->count());
         self::assertSame(0, $crawler->filter('textarea.form__result-html')->count());
         self::assertSame(1, $crawler->filter('html div.error-message', 'Unfortunately, we can\'t fix what doesn\'t exist ! Please enter something to fix.')->count());
